Extract logger out into a package
