+++ app.go
@@ -2,13 +2,9 @@ package main
 
 import (
 	"context"
-	"fmt"
-	"net/url"
-	"os"
-	"runtime"
-	"time"
 
 	"github.com/adrg/xdg"
+	"github.com/marcus-crane/october/pkg/logger"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 
@@ -28,15 +24,6 @@ type App struct {
 	KoboService *KoboService
 }
 
-// newWinFileSink creates a log sink on Windows machines as zap, by default,
-// doesn't support Windows paths. A workaround is to create a fake winfile
-// scheme and register it with zap instead. This workaround is taken from
-// the Github issue at https://github.com/uber-go/zap/issues/621
-func newWinFileSink(u *url.URL) (zap.Sink, error) {
-	// Remove leading slash left by url.Parse()
-	return os.OpenFile(u.Path[1:], os.O_WRONLY|os.O_APPEND|os.O_CREATE, 0644)
-}
-
 // NewApp creates a new App application struct
 func NewApp() (*App, error) {
 	configPath, err := xdg.ConfigFile(configFilename)
@@ -50,23 +37,7 @@ func NewApp() (*App, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to initialise settings")
 	}
-	logFile := fmt.Sprintf("october/logs/%s.log", time.Now().Format("2006-01-02"))
-	logPath, err := xdg.DataFile(logFile)
-	config := zap.NewProductionConfig()
-	if runtime.GOOS == "windows" {
-		err := zap.RegisterSink("winfile", newWinFileSink)
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to register winfile sink")
-		}
-		logPath = "winfile:///" + logPath
-	}
-	config.OutputPaths = []string{"stdout", logPath}
-	logger, err := config.Build()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to initialise logger")
-	}
-	defer logger.Sync()
-	sugaredLogger := logger.Sugar()
+	sugaredLogger := logger.Init()
 	app := &App{
 		settings: loadedSettings,
 		logger:   sugaredLogger,
